refactor: 관심사 api 훅 분리
