+++ src/api/hooks/index.ts
@@ -1,12 +1,12 @@
 export * from './use-assign-quest'
 export * from './use-repeat-quest'
 export * from './use-history'
-export * from './use-interest-graph'
-export * from './use-interest'
 export * from './use-major-interest'
 export * from './use-profile'
 export * from './use-quest-setting'
 export * from './use-streak'
+
+export * from './interest'
 
 export * from './api-mutate'
 export * from './error-handler'
